# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for managing players.",
        "title": "Players API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/health": {
            "get": {
                "description": "Check the health of the service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "Service unavailable",
                        "schema": {}
                    }
                }
            }
        },
        "/players": {
            "get": {
                "description": "Get a list of all players",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Get all players",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Player"
                            }
                        }
                    }
                }
            }
        },
        "/players/range": {
            "get": {
                "description": "Get a range of players with an offset",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Get a range of players",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Player"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/players/{id}": {
            "get": {
                "description": "Get a player by their ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Get player by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Player"
                        }
                    },
                    "400": {
                        "description": "Invalid player ID",
                        "schema": {}
                    },
                    "404": {
                        "description": "Player not found",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "main.Player": {
            "type": "object",
            "properties": {
                "bats": {
                    "type": "string"
                },
                "bbrefID": {
                    "type": "string"
                },
                "birthCity": {
                    "type": "string"
                },
                "birthCountry": {
                    "type": "string"
                },
                "birthDay": {
                    "type": "integer"
                },
                "birthMonth": {
                    "type": "integer"
                },
                "birthState": {
                    "type": "string"
                },
                "birthYear": {
                    "type": "integer"
                },
                "deathCity": {
                    "type": "string"
                },
                "deathCountry": {
                    "type": "string"
                },
                "deathDay": {
                    "type": "integer"
                },
                "deathMonth": {
                    "type": "integer"
                },
                "deathState": {
                    "type": "string"
                },
                "deathYear": {
                    "type": "integer"
                },
                "debut": {
                    "type": "string"
                },
                "finalGame": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "nameFirst": {
                    "type": "string"
                },
                "nameGiven": {
                    "type": "string"
                },
                "nameLast": {
                    "type": "string"
                },
                "playerID": {
                    "type": "string"
                },
                "retroID": {
                    "type": "string"
                },
                "throws": {
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}